# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aftersales Api - SP21",
        "description": "List Aftersales Api",
        "version": "1.0"
    },
    "paths": {
        "/feedbacks": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Show list feedbacks",
                "operationId": "06ef48a76c0e8cc754a917b101dd574d",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": 2
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "example content"
                                                    },
                                                    "created_at": {
                                                        "type": "timestamp",
                                                        "example": "2022-01-20T09:06:33.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "timestampl",
                                                        "example": "2022-01-20T09:06:33.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            }
        },
        "/feedbacks/{id}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Show a feedback with certain $id",
                "operationId": "6496919e6451e48076284d8703da18d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "created_at": {
                                            "type": "timestamp",
                                            "example": "2022-01-20T09:06:33.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "timestamp",
                                            "example": "2022-01-20T09:06:33.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "API key is missing."
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}